+++ server/feed2/src/main/scala/info/glennengstrand/news/Participant.scala
@@ -20,20 +20,20 @@ object Participant {
       List("name")
     }
     def upsertOutputs: Iterable[(String, String)] = {
-      List(("id", "Long"))
+      List(("id", "Int"))
     }
     def fetchOrder: Map[String, String] = {
       Map()
     }
   }
   val bindings = new ParticipantBindings
-  def create(id: Long, name: String): Participant = {
+  def create(id: Int, name: String): Participant = {
     IO.settings.getProperty(IO.jdbcVendor) match {
       case "mysql" => new Participant(id, name) with MySqlWriter with RedisCacheAware
       case _ => new Participant(id, name) with PostgreSqlWriter with RedisCacheAware
     }
   }
-  def apply(id: Long) : Participant = {
+  def apply(id: Int) : Participant = {
     val criteria: Map[String, Any] = Map("id" -> id)
     val state: Map[String, Any] = IO.cacheAwareRead(bindings, criteria, reader, cache).head
     create(id, state("Moniker").asInstanceOf[String])
@@ -41,17 +41,17 @@ object Participant {
   def apply(state: String): Participant = {
     val s = IO.fromFormPost(state)
     val id = s.contains("id") match {
-      case true => s("id").asInstanceOf[String].toLong
-      case _ => 0l
+      case true => s("id").asInstanceOf[String].toInt
+      case _ => 0
     }
     create(id, s("name").asInstanceOf[String])
   }
 }
 
-case class ParticipantState(id: Long, name: String)
+case class ParticipantState(id: Int, name: String)
 
 /** represents a participant who may have friends and news feed */
-class Participant(id: Long, name: String) extends ParticipantState(id, name) with MicroServiceSerializable {
+class Participant(id: Int, name: String) extends ParticipantState(id, name) with MicroServiceSerializable {
   this: PersistentRelationalDataStoreWriter with CacheAware =>
 
   /** save participant state to db */
@@ -63,7 +63,7 @@ class Participant(id: Long, name: String) extends ParticipantState(id, name) wit
       "id" -> id
     )
     val result = write(Participant.bindings, state, criteria)
-    val newId = result.getOrElse("id", 0l).asInstanceOf[Long]
+    val newId = result.getOrElse("id", 0l).asInstanceOf[Int]
     invalidate(Participant.bindings, criteria)
     Participant.create(newId, name)
   }
